Add token check for delete mutations

+++ graphql/resolvers/student/deleteStudentResolver.js
@@ -1,5 +1,9 @@
 const models = require("../../../models");
-module.exports = (source, { id }) => {
+module.exports = (source, { id }, { tokenPayload }) => {
+  if(!tokenPayload) {
+    return null;
+  }
+
   return models.Student.destroy({
     where: {
       id,
